requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Machine Learning & AI
transformers==4.53.1
torch==2.7.1
accelerate==0.25.0

# Model Dependencies
sentencepiece==0.1.99
protobuf==4.25.1

# HTTP & Utilities
requests>=2.31.0
python-dotenv==1.0.0

# Development & Testing (optional)
# pytest>=7.0.0
# pytest-flask>=1.2.0

# Performance & Monitoring (optional)
# numpy>=1.17
# scipy>=1.7.0
# psutil>=5.8.0 